requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==3.2.0
certifi==2025.1.31
cffi==2.0.0
click==8.1.8
distro==1.9.0
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
openai==1.69.0
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.1
pydantic_core==2.33.0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
uvicorn==0.34.0